+++ database/migrations/2021_01_27_203803_create_black_lists_table.php
@@ -15,8 +15,8 @@ class CreateBlackListsTable extends Migration
     {
         Schema::create('black_lists', function (Blueprint $table) {
             $table->id();
-            $table->integer('user_id')->references('id')->on('users');
-            $table->integer('blocked_id')->references('id')->on('users');
+            $table->bigIncrements('user_id')->references('id')->on('users');
+            $table->bigIncrements('blocked_id')->references('id')->on('users');
             $table->timestamps();
         });
     }
